Introducing "Processes". DebugGrid is now a Process
A "Process" is a generic piece of dynamically loaded code that has an update() function called every frame.
It is intended to be used for higher level logic than a "Module", which are statically loaded. Where a Module
might be the CameraSystem or ResourceManager, a Process might be an Actor or Event. They are not critical to the engine, so it can initialize (and run) without them.